Implement xs_cpy function
The xs_cpy function copies the whole content from source to destination.
The copy on write technique is implemented, meaning that rather than
allocating a new memory space and copying the string from source, the
string ptr in destination actually points to same memory address as the
source string ptr. When someone wants to modify a object and the related
string data is shared with at least two objects, first a new memory
space will be allocated and initialized with identical content, then the
modifications can be done to the new memory space. In order to use the
copy on write technique, memory is always allocated with 4 additional
bytes to store a reference count (refcnt), and refcnt is always located
at the leading 4 bytes of each allocated memory.